refactor(navigation): remove unused route to navigation conversion and provider
