+++ apps/frontend/src/app/app.routes.ts
@@ -10,17 +10,14 @@ export const appRoutes: Route[] = [
   {
     path: 'checkout',
     component: CheckoutComponent,
-    data: { pageTitle: 'Checkout', showInMenu: false },
   },
   {
     path: 'orders',
     loadChildren: () => import('@demo-shop/order').then(m => m.orderRoutes),
-    data: { pageTitle: 'Orders', showInMenu: true },
   },
   {
     path: 'products',
     loadChildren: () => import('@demo-shop/product').then(m => m.productRoutes),
-    data: { pageTitle: 'Products', showInMenu: true },
   },
   {
     path: '**',
