+++ libs/navigation/src/index.ts
@@ -1,2 +1,1 @@
 export * from './lib/components/navigation/navigation.component';
-export * from './lib/util/provide-navigation';
