+++ libs/navigation/src/lib/enums/navigation-type.ts
@@ -1,3 +1,9 @@
+/**
+ * Enum defining the types of navigation items in the application
+ */
 export enum NavigationType {
+  /**
+   * Represents a navigational item that is an Angular route
+   */
   ROUTE = 'ROUTE',
 }
