+++ libs/navigation/src/lib/services/navigation.service.ts
@@ -1,14 +1,21 @@
 import { inject, Injectable } from '@angular/core';
 import { AuthFacade, PermissionStrategy } from '@demo-shop/auth';
-import { NavigationItem, RouteItem } from '../models/navigation-item';
+import { NavigationItem } from '../models/navigation-item';
 import { NavigationType } from '../enums/navigation-type';
+import { RouteItem } from '../models/route-item';
 
+/**
+ * Service responsible for managing application navigation items
+ */
 @Injectable({
   providedIn: 'root',
 })
 export class NavigationService {
   readonly #authFacade = inject(AuthFacade);
 
+  /**
+   * Collection of all available navigation items in the application
+   */
   private readonly menuItems: NavigationItem[] = [
     new RouteItem('products', 101, {
       route: 'products',
@@ -19,6 +26,14 @@ export class NavigationService {
     }),
   ];
 
+  /**
+   * Returns navigation items filtered by type and user permissions
+   *
+   * This method filters the navigation items based on:
+   * 1. The specified navigation type
+   * 2. The user's permissions (if a permission strategy is defined)
+   * The resulting items are sorted by their order property in ascending order.
+   */
   getFilteredItems(type: NavigationType): NavigationItem[] {
     const isType = (itemType: NavigationType) => itemType === type;
     const allowDisplay = (strategy?: PermissionStrategy) => !strategy || this.#authFacade.hasPermission(strategy);
